Ionic Popovers, Loading, and Gestures
Added a popover, loading spinner, and delete gesture to favorites.html

+++ Ionic/conFusion/www/js/controllers.js
@@ -221,19 +221,29 @@ angular.module('conFusion.controllers', [])
             
                     }])
 
-        .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+        .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
             
             $scope.baseURL = baseURL;
             $scope.shouldShowDelete = false;
+            
+            $ionicLoading.show({
+                template: '<ion-spinner></ion-spinner> Loading...'
+            });
 
             $scope.favorites = favoriteFactory.getFavorites();
             
             $scope.dishes = menuFactory.getDishes().query(
                 function (response) {
                     $scope.dishes = response;
+                    $timeout(function() {
+                        $ionicLoading.hide();
+                    }, 1000);
                 },
                 function (response) {
                     $scope.message = "Error: " + response.status + " " + response.statusText;
+                    $timeout(function() {
+                        $ionicLoading.hide();
+                    }, 1000);
                 });
     
             console.log($scope.dishes, $scope.favorites);
@@ -244,7 +254,18 @@ angular.module('conFusion.controllers', [])
             }
 
             $scope.deleteFavorite = function (index) {
-                favoriteFactory.deleteFromFavorites(index);
+                
+                var confirmPopup = $ionicPopup.confirm({
+                    title: 'Confirm Delete',
+                    template: 'Are you sure you want to delete this favorite?'
+                });
+                
+                confirmPopup.then(function(res) {
+                    if (res) {
+                        console.log('OK TO DELETE');
+                        favoriteFactory.deleteFromFavorites(index);
+                    }
+                })
                 $scope.shouldShowDelete = false;
             }
         }])
